Added delete entity frontend logic

+++ frontend/src/assets/api/apiCalls.ts
@@ -1,7 +1,7 @@
 import axios, { AxiosResponse } from "axios"
 import { Entity, Product, ProductInstance } from "../../shared/types"
 import { getProductStageFromId, getProductLocationFromId, formatUnixTimestampToDatetime } from "../../utils/typeUtils"
-import { addCustomer, addDistributor, addManufacturer, addRetailer, getContractProductInfo, isManufacturer, produceProduct } from "../api/contractCalls"
+import { addCustomer, addDistributor, addManufacturer, addRetailer, getContractProductInfo, getEntityRole, isManufacturer, produceProduct } from "../api/contractCalls"
 
 export const api = axios.create({
 	baseURL: "http://localhost:3000/api",
@@ -100,11 +100,75 @@ export const createEntity = async ({
         console.error('Error creating entity:', error)
         throw error
     }
-};
+}
 
 export const deleteEntity = async ({ id }: { id: string }): Promise<AxiosResponse<{ message: string; data: Entity }>> => {
-	const res = await api.delete("/delete-entity", { params: { id } })
-	return res
+	try {
+		const entityRole = await getEntityRole()
+		let metamaskAddress
+		
+		switch (entityRole) {
+            case "manufacturer":
+                const manufacturerAccount = await addManufacturer()
+
+                if (manufacturerAccount) {
+					console.log('Manufacturer account:', manufacturerAccount)
+                    metamaskAddress = manufacturerAccount
+                } else {
+                    const error = console.error('Error adding manufacturer')
+					throw error
+                }
+				break
+            case "customer":
+				const customerAccount = await addCustomer()
+
+                if (customerAccount) {
+					console.log('Customer account:', customerAccount)
+                    metamaskAddress = customerAccount
+                } else {
+                    const error = console.error('Error adding customer')
+					throw error
+                }
+                break
+            case "retailer":
+				const retailerAccount = await addRetailer()
+
+                if (retailerAccount) {
+					console.log('Retailer account:', retailerAccount)
+                    metamaskAddress = retailerAccount
+                } else {
+                    const error = console.error('Error adding retailer')
+					throw error
+                }
+                break
+            case "distributor":
+                const distributorAccount = await addDistributor()
+
+                if (distributorAccount) {
+					console.log('Distributor account:', distributorAccount)
+                    metamaskAddress = distributorAccount
+                } else {
+                    const error = console.error('Error adding distributor')
+					throw error
+                }
+                break
+            default:
+				const error = console.error('Error adding entity. Specified role not valid.')
+				throw error
+        }
+
+		console.log(metamaskAddress)
+		if (metamaskAddress !== null) {
+			const res = await api.delete("/delete-entity", { params: { id } })
+			return res
+		} else {
+			const error = console.error("Error creating entity. Can't catch the contract event.")
+			throw error
+		}
+	} catch (error) {
+        console.error('Error deleting entity:', error)
+        throw error
+    }
 }
 
 export const addProduct = async ({ name, description }: { name: string; description: string }): Promise<AxiosResponse<{ message: string; data: Product }>> => {
